ISSUE #5310 - create and use useResizable
